feat: add forms.maxWidth token
- moves mathWithUnits util to eui-theme-common package to enable shared usage
