+++ packages/eui-theme-common/src/global_styling/functions/index.ts
@@ -7,3 +7,4 @@
  */
 
 export * from './size';
+export * from './math';
