+++ packages/eui-theme-common/src/global_styling/functions/math.test.ts
